Don't exit search mode after deleting a reading list
It is possible to delete a reading list while in search mode by first
selecting a list, then starting a search, then clicking the Remove
current reading list button. Exiting search mode and loading the next
selected reading list is arbitrary and unasked for in this case.

Updating history is still necessary to maintain the invariant
YACReaderHistoryController::currentSourceContainer().stillExists().