Store QPersistentModelIndex in YACReaderLibrarySourceContainer
Quote from the documentation for QModelIndex:
Note: Model indexes should be used immediately and then discarded. You
should not rely on indexes to remain valid after calling model functions
that change the structure of the model or delete items. If you need to
keep a model index over time use a QPersistentModelIndex.

QModelIndex YACReaderLibrarySourceContainer::sourceModelIndex can become
invalid because of a folder update or an addition/removal of a fodler or
list. Navigating back or forward in history to such an invalidated index
resulted in either a crash or selection/loading of a wrong index. With
this commit invalidated indices are removed from history instead.

In addition, LibraryWindow::reloadAfterCopyMove(), which is called after
a folder is updated or after a comic is drag&dropped into a folder,
properly refreshes the currently loaded ComicsView page now. The
(getCurrentFolderIndex() == mi) condition for reloading at the top of
this function does not make sense in search mode or when a reading list
is loaded in ComicsView. This condition is not reliable even if a folder
is loaded, because a parent folder could have been updated. The new
implementation makes sure that the currently loaded view is always
up-to-date.