LibraryWindow::rename: use currentSourceModelIndex()
LibraryWindow::rename() checks if the current foldersView index is valid
to determine whether the current library name is used as the toolbar
title. The result of this check is sometimes incorrect. For example,
renaming a library when the root folder is current and a reading list is
loaded changes the title from the loaded list's name to the new library
name. A similar unjustified title replacement occurs in search mode.

Checking if currentSourceModelIndex() is valid instead should always
produce the correct result. Add an assertion to detect out of sync
toolbar title.