Update status/type of selected rather than current folder
FolderModel::updateFolderCompletedStatus(),
FolderModel::updateFolderFinishedStatus(),
FolderModel::updateFolderManga() unconditionally dereference
internalPointer() of each index in their QModelIndexList argument.
Therefore these functions assume that all indices in the list are valid
(invalid/root index's internalPointer() is nullptr). Explicitly assert
this reasonable assumption.

When the current foldersView's index is the root index, triggering any
of the 6 status/type update actions passes an invalid index to
aforementioned FolderModel's member functions, which causes a crash.
Show the existing fitting "No folder selected" message instead of
passing an invalid index and crashing.

Updating status/type of the selected rather than the current folder is
more intuitive because the current folder is not highlighted in the UI.
Note that the root index is not necessarily current when no folder is
selected as the user can deselect a folder while keeping it current by
selecting a reading list or clicking outside of all folders in
foldersView. The top (not root) folder can be current in search mode as
Qt code sometimes ensures a valid current index.