Show status and type actions applicable to selected folder
The actions in foldersView's context menu are applied to the selected
folder. Therefore actions applicable to that folder should be put in the
context menu rather than those applicable to the folder at the
right-click position. Usually these two folders are the same, but not
always. For instance, it is possible to right-click outside of all
folders in foldersView while some folder is selected. Folder indices can
also be mixed up when the context menu is invoked in search mode.

Don't show status and type actions at all when no folder is selected as
triggering them in this case changes nothing and asks the user to select
a folder first. Extract code that conditionally adds these options into
a helper member function to reduce nesting. Note that even though the
context menu no longer contains status and type actions when no folder
is selected, the "No folder selected" message can still be shown when a
status or type update action is triggered via a shortcut.