Reflect folder status change via a shortcut in the UI immediately
When a folder is set as uncompleted/completed/read/unread via a
shortcut, the folder status indicator in the UI is not updated until
the affected index's row or the entire foldersView's viewport is updated
for another reason. Emitting QAbstractItemModel::dataChanged() signal
with correct arguments fixes this issue for the read/unread status. In
addition, YACReaderFoldersView::visualRect(index) has to be widened to
fix the issue for the uncompleted/completed status. Even though manga
folder type is not reflected in the UI, emit dataChanged() with
corresponding arguments from FolderModel::updateFolderManga() for
consistency.

An alternative to widening visualRect(index) is to move the
QStyledItemDelegate::paint() call to the top of
YACReaderFoldersViewItemDeletegate::paint() and draw the uncompleted
indicator inside option.rect (which equals the default visualRect for
the rendered index). For example, x=option.rect.x() rather than x=0 can
be passed to QPainter::drawRect(). But this would move the vertical line
to the right - just left of the folder icon, which may be undesirable.