ReadingListModel::dropSublist: drop into correct position
When a sublist is dragged and dropped below its current position and not
into another sublist, it ends up one row lower than the drop position.
When a sublist is dropped at the end of its parent reading list, destRow
becomes out of bounds and the Qt warning "QList::insert(): Index out of
range." is printed. This warning is replaced by an assertion in Qt 6.

The drag&drop behavior when a sublist is dropped inside another sublist
is convenient and not incorrect. Therefore this commit preserves it.

Call beginMoveRows() and endMoveRows(). When these two member functions
are not called, sublist navigation history is messed up because
YACReaderHistoryController's QPersistentModelIndex sources are not
adjusted. Namely, the behavior is as if integer indices are stored in
the navigation history rather than persistent indices.

beginMoveRows() returns false when destRow equals sourceRow or
sourceRow + 1. This is the correct condition for the early return of
false as no sublist reordering takes place.